+++ rotkehlchen/externalapis/utils.py
@@ -5,6 +5,7 @@ from hexbytes import HexBytes as Web3HexBytes
 from rotkehlchen.api.websockets.typedefs import WSMessageType
 from rotkehlchen.chain.binance_sc.constants import BINANCE_SC_GENESIS
 from rotkehlchen.chain.hyperliquid.constants import HYPERLIQUID_GENESIS
+from rotkehlchen.chain.ink.constants import INK_GENESIS
 from rotkehlchen.chain.monad.constants import MONAD_GENESIS
 from rotkehlchen.chain.robinhood.constants import ROBINHOOD_GENESIS
 from rotkehlchen.chain.scroll.constants import SCROLL_GENESIS
@@ -92,6 +93,8 @@ def get_earliest_ts(chain_id: SUPPORTED_CHAIN_IDS) -> Timestamp:
             return SONIC_GENESIS
         case ChainID.ROBINHOOD:
             return ROBINHOOD_GENESIS
+        case ChainID.INK:
+            return INK_GENESIS
         case ChainID.POLYGON_POS:
             return Timestamp(1590856200)